[CoreBundle] exposes copy to clipboard as ES6 module
fixes webpack
